Lab2-RoutingAgent_RAG/iphone_assistant/requirements.txt
streamlit
snowflake-connector-python
python-dotenv